% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransCLIME.R
\name{Agg}
\alias{Agg}
\title{LS aggregation function}
\usage{
Agg(Theta.init, X.til)
}
\arguments{
\item{Theta.init}{is a column-bound combination of Theta.cl and Theta.hat}

\item{X.til}{is a subset of primary samples}
}
\value{
is SOMETHING
}
\description{
LS aggregation function
}
