% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransCLIME.R
\name{cv.clime}
\alias{cv.clime}
\title{Cross validation for selecting tuning parameters}
\usage{
cv.clime(X, nfold = 5)
}
\arguments{
\item{X}{is SOMETHING}

\item{nfold}{is the number of folds (default: 5)}
}
\value{
is SOMETHING
}
\description{
Cross validation for selecting tuning parameters
}
