% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransCLIME.R
\name{DB.clime.FDR}
\alias{DB.clime.FDR}
\title{DB.clime.FDR is SOMETHING}
\usage{
DB.clime.FDR(Theta, X)
}
\arguments{
\item{Theta}{is SOMETHING}

\item{X}{is SOMETHING}
}
\value{
is SOMETHING
}
\description{
DB.clime.FDR is SOMETHING
}
