% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransCLIME.R
\name{Dist}
\alias{Dist}
\title{Compute the estimation errors based on the test samples}
\usage{
Dist(X.test, Theta.hat, Theta0)
}
\arguments{
\item{X.test}{is the set of test samples}

\item{Theta.hat}{is SOMETHING}

\item{Theta0}{is SOMETHING}
}
\value{
is a named list containing \itemize{
\item \code{Frob} is SOMETHING
\item \code{S} is SOMETHING
\item \code{te} is SOMETHING
}
}
\description{
Compute the estimation errors based on the test samples
}
