% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransCLIME.R
\name{jgl.fun}
\alias{jgl.fun}
\title{Joint graph estimator function from Guo et al. (2011)}
\usage{
jgl.fun(X.all, n.vec, lam.const = NULL)
}
\arguments{
\item{X.all}{is SOMETHING}

\item{n.vec}{is SOMETHING}

\item{lam.const}{is SOMETHING (default: NULL)}
}
\value{
is a named list containing \itemize{
\item \code{Theta.hat} is SOMETHING
\item \code{lam.const} is SOMETHING
}
}
\description{
Joint graph estimator function from Guo et al. (2011)
}
\seealso{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3412604/}
}
