% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransCLIME.R
\name{Myfastclime.s}
\alias{Myfastclime.s}
\title{Algorithm based on fastclime package
min \|Theta\|_1
subject to \|cov(X)\%\emph{\%Theta-Bmat\|_max <=lambda (if X is the raw sample)
subject to \|X\%}\%Theta-Bmat\|_max <=lambda (if X is the sample covariance matrix)}
\usage{
Myfastclime.s(X, Bmat, lambda = 0.1, scale = T, n)
}
\arguments{
\item{X}{is SOMETHING}

\item{Bmat}{is SOMETHING}

\item{lambda}{is SOMETHING (default: 0.1)}

\item{scale}{is SOMETHING (default: T)}

\item{n}{is SOMETHING}
}
\value{
is a named list containing \itemize{
\item \code{Theta.hat} SOMETHING
\item \code{conv} SOMETHING
}
}
\description{
Algorithm based on fastclime package
min \|Theta\|_1
subject to \|cov(X)\%\emph{\%Theta-Bmat\|_max <=lambda (if X is the raw sample)
subject to \|X\%}\%Theta-Bmat\|_max <=lambda (if X is the sample covariance matrix)
}
