% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransCLIME.R
\name{Spd.proj}
\alias{Spd.proj}
\title{Compute the semi-positive definite projection of SigA.hat
with smallest eigenvalue lower bounded by eps}
\usage{
Spd.proj(SigA.hat, eps = NULL)
}
\arguments{
\item{SigA.hat}{is the input matrix}

\item{eps}{is the epsilon value}
}
\value{
is a named list containing \itemize{
\item \code{mat} SOMETHING
\item \code{conv} SOMETHING
}
}
\description{
Compute the semi-positive definite projection of SigA.hat
with smallest eigenvalue lower bounded by eps
}
