% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransCLIME.R
\name{Trans.CLIME}
\alias{Trans.CLIME}
\title{The main TransCLIME algorithm}
\usage{
Trans.CLIME(X, X.A, const, agg = T, X.til = NULL, Theta.cl)
}
\arguments{
\item{X}{is the primary data set}

\item{X.A}{is the set of X^{(k)} for k in A}

\item{const}{is SOMETHING}

\item{agg}{is a boolean to perform LS aggregation or not (default: T)}

\item{X.til}{is the sample set for aggregation (default: NULL)}

\item{Theta.cl}{is the CLIME estimator}
}
\value{
is SOMETHING
}
\description{
The main TransCLIME algorithm
}
