% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransCLIME.R
\name{l1proj}
\alias{l1proj}
\title{Sourced from: https://rdrr.io/github/celiaescribe/BDcocolasso/src/R/ADMM_proj.R}
\usage{
l1proj(v, b)
}
\arguments{
\item{v}{is SOMETHING}

\item{b}{is SOMETHING}
}
\value{
is SOMETHING
}
\description{
Sourced from: https://rdrr.io/github/celiaescribe/BDcocolasso/src/R/ADMM_proj.R
}
